requirements.txt
tqdm
rotary_embedding_torch
transformers==4.19
tokenizers
inflect
progressbar
einops
unidecode
scipy
librosa
numpy==1.20.0
numba==0.48.0
torchaudio
threadpoolctl
llvmlite
appdirs
sanic==22.6.2